Added a singleton check so only one instance of the program runs at a time
This is achieved with a named mutex that's created on boot.  If the
mutex already exists then the implication is that another instance of
the program is already running and the current instance exits.
Otherwise the program works like normal.  This avoids the pitfalls of a
pid file since it won't persist in the event of an unforseen crash.